+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## [Unreleased]
 
+- Userinfo endpoint now returns `aud` claim.
+
 ## [0.1.0] - 2023-02-25
 
 - Initial release